+++ src/application/domain/lokasi/lokasi.domain.js
@@ -1,4 +1,5 @@
 const Lokasi = require('../../../model/lokasi');
+const Barang = require('../../../model/barang');
 
 async function searchLokasi(namaLokasi){
     const response = await Lokasi.find({namaLokasi: namaLokasi});
@@ -101,17 +102,24 @@ async function updateLokasi(id, dataLokasi) {
 }
 async function deleteLokasi(id) {
     try {
-        await Lokasi.deleteOne({
-            _id: id
+        const barang = await Barang.find({
+            lokasi: id
         });
-        return {
-            status: 200,
-            message: 'Berhasil menghapus lokasi penyimpanan'
-        };
+        if(!barang){
+            await Lokasi.deleteOne({
+                _id: id
+            });
+            return {
+                status: 200,
+                message: 'Berhasil menghapus lokasi penyimpanan'
+            };
+        }else{
+            throw new Error(`Tidak dapat di hapus. ${barang.namaBrang} sedang disimpan di tempat penyimpanan ini.`);
+        }
     } catch (error) {
         return {
             status: 500,
-            message: 'Gagal menghapus lokasi penyimpanan'
+            message: error.message
         };
     }
 }
